Developing for Win32 using Delphi (Object Pascal)
Seem to have misplaced the original document, this is the one which
I had exported (some issues with hierarchy, need to fix later) to
html for my website.

NOTE: This is before I had got to my senses and started writing the
html documents directly in vim and or using lyx as word processor.

NOTE: Done sometime around 1998