I18N and L10N for applications in Linux
During IT.com 2000 in karnataka, We from blug/ilug wanted to help
popularise linux usage in India, by also showing that it can be
used in ones native language, If that is what one prefers.

So I looked into the steps required to achieve the same and created this
note to give a rough idea of the things/steps involved, to help volunteers
as well as people who may be interested in localised linux.

As part of the same I had also created a onscreen keyboard, as well as
made a icq msging app support multiple languages in the messages by adding
tagged messages support, ...