Linux the Kernel and OS, Some of its low level Intricacies
THis is a potentially slightly different version of what I had posted
to linux kernel mailing list, as a preliminary linux porting guide
draft.

And I was lucky to get feedback from Alan Cox at that time, that
these could be the basis for a set of good gearheads articles.
But due to work and life commitments at that time couldnt take it
further.

I think it was also later referenced from a linux mips porting guide.

- 2020 -