Changes that came with Linux kernel 2.6
Come to think of it 2.6 had a good set of changes which made linux
better in many ways.

I had prepared this for a presentation at the local blug, for the
linux birthday, as well as for self reference. Rather due to lack
of time, I think, this itself was the presentation.

It was also the march towards the namespaces and containerisation
we see today, which I had noted then during the presentation.