mutex_ks - a sync mechanism for uni/smp systems
When checking out the still experimental 2.6.0 kernel test release,
found a issue with my cd drive. Inturn found it to be a issue related
to the sync primitives in kernel, their flexibilities (or the lack there
off) and the way they were used.

So had created this sync mechanism which selects between the existing
mechanisms based on uni processor or smp system, in order to provide
a mutual exclusion, without enforcing exclusive execution be it SMP
or Uni processor machine. It solved the issue on my machine at the same
time.

I think I posted this also to the linux kernel mailing list, but didnt
persist or follow up.

Those were the days, when a common man could only dream of SMP, unless
they were willing to starve. And technology progress is a wonderful
thing, now chances are even a cheap throw away device may have multiple
processing logics in them either in symmetric or asymmetric way.

_2020_