Notes for using KiCad for Schematics and PCB
Provides some quick notes for using KiCad for Schematics and PCB.

This is a capture of what I did the last time I used KiCad.

NOTE: As things keep revolving in and out of existance in my univ,
this is what I remembered/thought and did and not necessarily the
optimal thing, as I tend to remember things only partially and then
jump into taking it from there, as it comes. RTFM only occurs once
in a bluemoon, usually, or for that matter thinking through things
fully when experimenting with short windows, so take these with a
pinch of salt.