Some basicNotes on interfacingPeripherals to a MCU
Some basic notes wrt interfacing peripherals to MCU, created as a
quick guide for teams which may be participating in the

Swadeshi Chips initiative being conducted by Indian Govt, IIT and CDAC.

This tries to answer some of the queries which I noticed in the forum
created for this initiative.

Will also act as a quick refresh for myself if required in future.